Complete_Places_Server
Added initializing function (prefixmatch.h)
