+++ place_svc.c
@@ -4,6 +4,7 @@
  */
 
 #include "place.h"
+#include "prefixmatch.h"
 #include <stdio.h>
 #include <stdlib.h>
 #include <rpc/pmap_clnt.h>
@@ -11,13 +12,190 @@
 #include <memory.h>
 #include <sys/socket.h>
 #include <netinet/in.h>
+#include <iostream>
+#include <sstream>
+#include <cstdlib>
+#include <cstring>
+#include <stack>
+#include <algorithm>
 
 using namespace std;
 
 #ifndef SIG_PF
 #define SIG_PF void(*)(int)
 #endif
+trie::trie() 
+{ 
+    root = new_node(0);
+}
+trie trie::instance;
+node* trie::new_node(int data) 
+{   
+    node *Q = new node; 
+    Q->data = data;
+    Q->isLeaf = false;
+    Q->isCity = false;
+    Q->count = 0;
+    return Q; 
+}
+void trie::removeSpace(char *str) {
+  char *p1 = str, *p2 = str;
+  do 
+    while (*p2 == ' ')
+      p2++;
+  while (*p1++ = *p2++);
+}
+void trie::uppercase(char* str) {
+    char c;
+    int i =0;
+    while(str[i]){
+        str[i] = toupper(str[i]);
+        i++;
+    }
+}
 
+void trie::add(char* S, PlacesInfo info) 
+{ 
+    node *cur = root;
+    int temp;
+    int size = strlen(S);
+    
+    for (int i = 0; i <= size; i++)
+    {
+        temp = S[i] - 'A';
+        if (!cur->child[temp]) {
+            cur->child[temp] = new_node(S[i]);
+        }
+        if(i!=size){
+            cur = cur->child[temp];
+            cur->count++;
+        }
+    }
+    cur->isCity = true;
+    temp = cur->count;
+    cur->info[temp].lat = info.lat;
+    cur->info[temp].lon = info.lon;
+    strncpy(cur->info[temp].state, info.state,2);
+    strcpy(cur->info[temp].city, S);
+    //cout<<S<<endl;
+}
+
+int trie::check(node *root, char *city, char* state, PlacesInfo& info) 
+{ 
+    node *cur = root;
+    node *nodeCheck;
+    
+    int count = 0;
+    int stateCount = 0;
+    int checkCount = 0;
+    int temp, status = 1;
+    int size = strlen(city);
+    while(count < size)
+    {
+        temp = city[count] - 'A';
+        if(cur->child[temp])
+        {
+            count++;
+            cur = cur->child[temp];
+        }
+        else
+            return 1;  
+    }
+
+    if(cur->isCity)
+    {
+        while(stateCount <= cur->count){
+            if(strncmp(state,cur->info[stateCount].state,2)==0)
+            {
+                status = 0;
+                info.lat = cur->info[stateCount].lat;
+                info.lon = cur->info[stateCount].lon;
+                strncpy(info.state,cur->info[stateCount].state,2);
+                strcpy(info.city,cur->info[stateCount].city);
+                break;
+            }
+            else{
+                status = 1;
+                stateCount++;
+            }
+        }
+    } 
+    else
+    {
+        while(cur->count==1&&!cur->isCity)
+        {
+            nodeCheck = cur->child[checkCount];
+            while(nodeCheck ==0)
+            {
+
+                nodeCheck = cur->child[checkCount];
+                checkCount++;
+            }
+            cur = nodeCheck;
+            checkCount = 0;
+        }
+        if(cur->isCity)
+        {
+            while(stateCount <= cur->count){
+                if(strncmp(state,cur->info[stateCount].state,2)==0)
+                {
+                    status = 0;
+                    info.lat = cur->info[stateCount].lat;
+                    info.lon = cur->info[stateCount].lon;
+                    strncpy(info.state,cur->info[stateCount].state,2);
+                    strcpy(info.city,cur->info[stateCount].city);
+                    break;
+                }
+                else{
+                    status = 1;
+                    stateCount++;
+                }
+            }
+        }
+        else
+            status = 2;
+    }
+    return status;
+}
+
+int trie::checkroot(char* input, char* state, PlacesInfo& info) 
+{ 
+    if (root && sizeof(input) > 0 && input[0] >= 'A') 
+        return check(root, input, state, info); 
+    else 
+        return 1; 
+}
+void
+startup(trie* ptr)
+{
+    trie dict = *ptr;
+    
+    FILE *pFile = fopen("places2k.txt","r");
+    char city[CITYBUFFER];
+    char line [BUFFERSIZE];
+    char temp[10];
+    PlacesInfo city_details;
+    int j = 0;
+    while(fgets(line, BUFFERSIZE, pFile)!=NULL)
+    {
+        strncpy(city, line+9, CITYBUFFER);
+        dict.removeSpace(city);
+        strcpy(city_details.city,city);
+        strncpy(city_details.state, line, 2);
+        memset(temp,' ',10);
+        dict.uppercase(city);
+        strncpy(temp, line+143, 10);
+        city_details.lat = atof(temp);
+        memset(temp,' ',10);
+        strncpy(temp, line+153, 11);
+        city_details.lon = atof(temp);         
+        dict.add(city, city_details);
+        memset(city,' ',CITYBUFFER);
+        memset(temp,' ',10);
+        
+    }
+    fclose(pFile);
+}
 static void
 places_dirprog_1(struct svc_req *rqstp, register SVCXPRT *transp)
 {
@@ -63,7 +241,8 @@ int
 main (int argc, char **argv)
 {
 	register SVCXPRT *transp;
-
+	trie *prefixTree = trie::getInstance();
+	startup(prefixTree);
 	pmap_unset (PLACES_DIRPROG, PLACES_DIR_VERS);
 
 	transp = svcudp_create(RPC_ANYSOCK);
